+++ plugins/memory/src/tests/wrappers.test.ts
@@ -1,6 +1,6 @@
 import { CacheDbPlugin, RetryDbPlugin } from '@routier/core/plugins';
 import { uuidv4 } from '@routier/core';
-import { describePluginContract, describeVectorSearch } from '@routier/test-utils';
+import { describePluginContract, describeVectorSearch, RENAMED_CALL_SELECTOR_TESTS } from '@routier/test-utils';
 import { MemoryPlugin } from '../MemoryPlugin';
 
 /**
@@ -18,13 +18,13 @@ import { MemoryPlugin } from '../MemoryPlugin';
 describePluginContract(
     'memory behind RetryDbPlugin',
     () => new RetryDbPlugin(new MemoryPlugin(`retry-${uuidv4()}`)),
-    { supportsRichTypes: true },
+    { supportsRichTypes: true, knownFailing: RENAMED_CALL_SELECTOR_TESTS },
 );
 
 describePluginContract(
     'memory behind CacheDbPlugin',
     () => new CacheDbPlugin(new MemoryPlugin(`cache-${uuidv4()}`)),
-    { supportsRichTypes: true },
+    { supportsRichTypes: true, knownFailing: RENAMED_CALL_SELECTOR_TESTS },
 );
 
 describeVectorSearch(